+++ src/main/java/com/dare/cinema_booking_system/screenings/service/ScreeningsService.java
@@ -1,6 +1,5 @@
 package com.dare.cinema_booking_system.screenings.service;
 
-import com.dare.cinema_booking_system.movies.dto.MovieResponse;
 import com.dare.cinema_booking_system.movies.entity.MovieEntity;
 import com.dare.cinema_booking_system.movies.service.MovieService;
 import com.dare.cinema_booking_system.rooms.entity.CinemaRoomEntity;
@@ -13,6 +12,7 @@ import com.dare.cinema_booking_system.screenings.entity.ScreeningsEntity;
 import com.dare.cinema_booking_system.screenings.entity.TimeSlot;
 import com.dare.cinema_booking_system.screenings.exceptions.ScreeningNotFoundException;
 import com.dare.cinema_booking_system.screenings.exceptions.ScreeningSlotAlreadyBookedException;
+import com.dare.cinema_booking_system.screenings.exceptions.ScreeningUpdateNotPossibleException;
 import com.dare.cinema_booking_system.screenings.repository.ScreeningSeatRepository;
 import com.dare.cinema_booking_system.screenings.repository.ScreeningsRepository;
 import lombok.AllArgsConstructor;
@@ -29,6 +29,7 @@ import java.util.List;
 @Service
 @AllArgsConstructor
 public class ScreeningsService {
+
 	private final ScreeningsRepository screeningsRepository;
 	private final ScreeningSeatRepository screeningSeatRepository;
 	private final MovieService movieService;
@@ -45,40 +46,83 @@ public class ScreeningsService {
 	}
 
 	public ScreeningsResponse getScreeningById(Long id) {
-		ScreeningsEntity screenings = getScreeningEntity(id);
+		ScreeningsEntity screening = getScreeningEntity(id);
 		log.info("Getting Screening by {} ID", id);
-		return modelMapper.map(screenings, ScreeningsResponse.class);
+		return modelMapper.map(screening, ScreeningsResponse.class);
 
 	}
 
 	public void deleteScreeningById(Long id) {
-		ScreeningsEntity screenings = getScreeningEntity(id);
-		screeningsRepository.delete(screenings);
-		log.info("Deleted Screening with {} ID", id);
-	}
 
-	public ScreeningsResponse createScreenings(ScreeningsRequest screeningsRequest) {
-		MovieEntity movieForScreening = movieService.getMovieEntityById(screeningsRequest.getMovieId());
-		CinemaRoomEntity roomForScreening = cinemaRoomService.getRoomEntity(screeningsRequest.getRoomId());
-		LocalDate dateForScreening = screeningsRequest.getScreeningDate();
-		TimeSlot  timeSlotForScreening = screeningsRequest.getTimeSlot();
+		boolean reservations = validateScreeningUpdate(id);
 
-		boolean validateScreeningSpot =validateScreeningSpot(screeningsRequest);
+		if (!reservations) {
+			ScreeningsEntity screenings = getScreeningEntity(id);
+			screeningsRepository.delete(screenings);
+			log.info("Deleted Screening with {} ID", id);
 
-		if(!validateScreeningSpot) {
-		ScreeningsEntity newScreeningEntity = new ScreeningsEntity(roomForScreening.getId(), movieForScreening,dateForScreening,timeSlotForScreening);
-		screeningsRepository.save(newScreeningEntity);
-		createScreeningSeats(roomForScreening, newScreeningEntity);
-
-		log.info("Screening created successfully");
-		return modelMapper.map(newScreeningEntity, ScreeningsResponse.class);
-
-		} else{
-			log.warn("Screening spot not available");
-			throw new ScreeningSlotAlreadyBookedException(roomForScreening.getId(),dateForScreening, timeSlotForScreening);
+		} else {
+			throw new ScreeningUpdateNotPossibleException(id);
 		}
 	}
 
+	public ScreeningsResponse createScreenings(ScreeningsRequest screeningsRequest) {
+
+		MovieEntity movie = movieService.getMovieEntityById(screeningsRequest.getMovieId());
+		CinemaRoomEntity room = cinemaRoomService.getRoomEntity(screeningsRequest.getRoomId());
+		LocalDate date = screeningsRequest.getScreeningDate();
+		TimeSlot  timeSlot = screeningsRequest.getTimeSlot();
+
+		boolean spotReserved = validateScreeningSpot(screeningsRequest);
+
+		if(!spotReserved) {
+			ScreeningsEntity screening = new ScreeningsEntity(room.getId(), movie,date,timeSlot);
+			screeningsRepository.save(screening);
+			createScreeningSeats(room, screening);
+
+			log.info("Screening created successfully");
+			return modelMapper.map(screening, ScreeningsResponse.class);
+
+		} else{
+			log.warn("Screening spot not available");
+			throw new ScreeningSlotAlreadyBookedException(room.getId(),date, timeSlot);
+		}
+	}
+	public ScreeningsResponse updateScreenings(Long screeningId,ScreeningsRequest screeningsRequest) {
+
+		ScreeningsEntity toUpdate = getScreeningEntity(screeningId);
+		MovieEntity requestedMovie = movieService.getMovieEntityById(screeningsRequest.getMovieId());
+		LocalDate requestedDate = screeningsRequest.getScreeningDate();
+		CinemaRoomEntity  requestedRoom = cinemaRoomService.getRoomEntity(screeningsRequest.getRoomId());
+		TimeSlot requestedTime = screeningsRequest.getTimeSlot();
+
+		boolean hasReservation = validateScreeningUpdate(screeningId);
+
+		if(!hasReservation) {
+			boolean spotIsBooked = validateScreeningSpot(screeningsRequest);
+
+			if(!spotIsBooked) {
+				List<ScreeningSeatEntity> newUpdatedScreeningSeats = createScreeningSeats(requestedRoom, toUpdate);
+
+				toUpdate.setMovie(requestedMovie);
+				toUpdate.setScreeningDate(requestedDate);
+				toUpdate.setTimeSlot(requestedTime);
+				toUpdate.setCinemaRoomId(requestedRoom.getId());
+				toUpdate.setScreeningSeatEntities(newUpdatedScreeningSeats);
+
+				screeningsRepository.save(toUpdate);
+				log.info("Screening with id {} updated successfully", screeningId);
+				return modelMapper.map(toUpdate, ScreeningsResponse.class);
+			} else {
+				log.warn("Screening spot not available");
+				throw new ScreeningSlotAlreadyBookedException(requestedRoom.getId(),requestedDate, requestedTime);
+			}
+		} else {
+			throw new ScreeningUpdateNotPossibleException(screeningId);
+		}
+
+
+	}
 
 	//Helper
 	private ScreeningsEntity getScreeningEntity(Long screeningId) {
@@ -90,9 +134,17 @@ public class ScreeningsService {
 	}
 
 	private List<ScreeningSeatEntity> createScreeningSeats(CinemaRoomEntity cinemaRoom, ScreeningsEntity screeningsEntity) {
-		List<SeatEntity> seats = cinemaRoom.getSeats();
+		List<SeatEntity> cinemaSeats = cinemaRoom.getSeats();
 		List<ScreeningSeatEntity> screeningSeats = screeningsEntity.getScreeningSeatEntities();
-		for(SeatEntity seat : seats) {
+		if (!screeningSeats.isEmpty()) {
+			List<ScreeningSeatEntity> oldScreeningSeats = screeningSeatRepository.getScreeningSeatsByscreeningsEntity(screeningsEntity);
+			screeningSeatRepository.deleteAllInBatch(oldScreeningSeats);
+			screeningSeatRepository.flush();
+			screeningSeats.clear();
+			log.info("Existing screening seats cleared");
+		}
+
+		for(SeatEntity seat : cinemaSeats) {
 			screeningSeats.add(new ScreeningSeatEntity(screeningsEntity, seat));
 		}
 		log.info("Seats for screening with {} id created successfully", screeningsEntity.getId());
@@ -105,4 +157,9 @@ public class ScreeningsService {
 				(screeningsRequest.getRoomId(),screeningsRequest.getScreeningDate(),screeningsRequest.getTimeSlot());
 
 	}
+
+	private boolean validateScreeningUpdate(Long screeningId) {
+		return screeningSeatRepository.hasReservedOrSoldSeats(screeningId);
+	}
+
 }
